[ci:force] move search service in alfresco content
