+++ docs/core/interfaces/search-configuration.interface.md
@@ -4,7 +4,7 @@ Added: v2.1.0
 Status: Active
 ---
 
-# [Search Configuration interface](../../../lib/core/src/lib/interface/search-configuration.interface.ts "Defined in search-configuration.interface.ts")
+# [Search Configuration interface](lib/content-services/src/lib/search/interfaces/search-configuration.interface.ts "Defined in search-configuration.interface.ts")
 
 Provides fine control of parameters to a search.
