+++ lib/content-services/src/lib/search/components/search.component.ts
@@ -15,7 +15,8 @@
  * limitations under the License.
  */
 
-import { SearchService, SearchComponentInterface } from '@alfresco/adf-core';
+import { SearchService } from '../services/search.service';
+import { SearchComponentInterface } from '../interfaces/search-configuration.interface';
 import {
     AfterContentInit,
     Component,
