+++ lib/content-services/src/lib/search/interfaces/search-configuration.interface.ts
