+++ lib/core/src/lib/services/public-api.ts
@@ -30,11 +30,9 @@ export * from './favorites-api.service';
 export * from './nodes-api.service';
 export * from './people-content.service';
 export * from './people-process.service';
-export * from './search.service';
 export * from './shared-links-api.service';
 export * from './discovery-api.service';
 export * from './comment-process.service';
-export * from './search-configuration.service';
 export * from './comment-content.service';
 export * from './login-dialog.service';
 export * from './external-alfresco-api.service';
