+++ lib/process-services/src/lib/form/widgets/content-widget/attach-file-widget-dialog.component.ts
@@ -17,9 +17,9 @@
 
 import { Component, Inject, ViewEncapsulation, ViewChild } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { ExternalAlfrescoApiService, AlfrescoApiService, LoginDialogPanelComponent, SearchService, TranslationService, AuthenticationService } from '@alfresco/adf-core';
+import { ExternalAlfrescoApiService, AlfrescoApiService, LoginDialogPanelComponent, TranslationService, AuthenticationService } from '@alfresco/adf-core';
 import { AttachFileWidgetDialogComponentData } from './attach-file-widget-dialog-component.interface';
-import { DocumentListService, SitesService } from '@alfresco/adf-content-services';
+import { DocumentListService, SitesService, SearchService } from '@alfresco/adf-content-services';
 import { Node } from '@alfresco/js-api';
 
 @Component({
